Extract denial formatting and subject extraction from main.rs (-309 lines)
Move format_denial_for_user() into new denial.rs module and
extract_subject() into classify.rs where it belongs with the other
tool classification logic. Tests moved alongside the functions.
